Start moving our changes out of the core Staking code.
