+++ pallets/staking/Cargo.toml
@@ -60,7 +60,6 @@ std = [
 	"pallet-babe/std",
 	"sp-arithmetic/std",
 	"rand_chacha",
-	"frame-benchmarking",
 ]
 runtime-benchmarks = [
 	"frame-benchmarking",
