+++ pallets/staking/src/lib.rs
@@ -289,8 +289,11 @@ pub mod testing_utils;
 
 pub mod inflation;
 pub mod slashing;
-pub mod types;
 pub mod weights;
+
+pub mod permissioned;
+pub use permissioned::PolymeshConvertCurve;
+pub mod types;
 
 pub mod pallet;
 
@@ -886,8 +889,6 @@ pub trait EraPayout<Balance> {
         total_staked: Balance,
         total_issuance: Balance,
         era_duration_millis: u64,
-        max_inflated_issuance: Balance,
-        non_inflated_yearly_reward: Balance,
     ) -> (Balance, Balance);
 }
 
@@ -896,8 +897,6 @@ impl<Balance: Default> EraPayout<Balance> for () {
         _total_staked: Balance,
         _total_issuance: Balance,
         _era_duration_millis: u64,
-        _max_inflated_issuance: Balance,
-        _non_inflated_yearly_reward: Balance,
     ) -> (Balance, Balance) {
         (Default::default(), Default::default())
     }
@@ -905,26 +904,27 @@ impl<Balance: Default> EraPayout<Balance> for () {
 
 /// Adaptor to turn a `PiecewiseLinear` curve definition into an `EraPayout` impl, used for
 /// backwards compatibility.
-pub struct ConvertCurve<T>(sp_std::marker::PhantomData<T>);
+pub struct ConvertCurve<T, C>(sp_std::marker::PhantomData<(T, C)>);
 
-impl<Balance: AtLeast32BitUnsigned + Clone, T: Get<&'static PiecewiseLinear<'static>>>
-    EraPayout<Balance> for ConvertCurve<T>
+impl<
+        Balance: AtLeast32BitUnsigned + Clone,
+        T: Config<CurrencyBalance = Balance>,
+        C: Get<&'static PiecewiseLinear<'static>>,
+    > EraPayout<Balance> for ConvertCurve<T, C>
 {
     fn era_payout(
         total_staked: Balance,
         total_issuance: Balance,
         era_duration_millis: u64,
-        max_inflated_issuance: Balance,
-        non_inflated_yearly_reward: Balance,
     ) -> (Balance, Balance) {
         let (validator_payout, max_payout) = inflation::compute_total_payout(
-            T::get(),
+            C::get(),
             total_staked,
             total_issuance,
             // Duration of era; more than u64::MAX is rewarded as u64::MAX.
             era_duration_millis,
-            max_inflated_issuance,
-            non_inflated_yearly_reward,
+            T::MaxVariableInflationTotalIssuance::get(),
+            T::FixedYearlyReward::get(),
         );
         let rest = max_payout.saturating_sub(validator_payout.clone());
         (validator_payout, rest)
