+++ pallets/staking/src/pallet/impls.rs
@@ -53,15 +53,12 @@ use super::{pallet::*, STAKING_ID};
 // Polymesh change
 // -----------------------------------------------------------------
 
-use frame_support::traits::schedule::Anon;
-use frame_support::traits::schedule::{DispatchTime, HIGHEST_PRIORITY};
 use frame_support::traits::DefensiveSaturating;
 
 use polymesh_primitives::traits::IdentityFnTrait;
-use polymesh_primitives::IdentityId;
-use polymesh_primitives::GC_DID;
 
 use crate::pallet::SlashingSwitch;
+use crate::types::PermissionedStaking;
 use crate::UnlockChunk;
 // -----------------------------------------------------------------
 
@@ -152,7 +149,7 @@ impl<T: Config> Pallet<T> {
         Ok(used_weight)
     }
 
-    pub(super) fn do_payout_stakers(
+    pub fn do_payout_stakers(
         validator_stash: T::AccountId,
         era: EraIndex,
     ) -> DispatchResultWithPostInfo {
@@ -309,7 +306,7 @@ impl<T: Config> Pallet<T> {
     /// Update the ledger for a controller.
     ///
     /// This will also update the stash lock.
-    pub(crate) fn update_ledger(controller: &T::AccountId, ledger: &StakingLedger<T>) {
+    pub fn update_ledger(controller: &T::AccountId, ledger: &StakingLedger<T>) {
         T::Currency::set_lock(
             STAKING_ID,
             &ledger.stash,
@@ -320,7 +317,7 @@ impl<T: Config> Pallet<T> {
     }
 
     /// Chill a stash account.
-    pub(crate) fn chill_stash(stash: &T::AccountId) {
+    pub fn chill_stash(stash: &T::AccountId) {
         // Polymesh change
         // -----------------------------------------------------------------
         Self::release_running_validator(stash);
@@ -496,51 +493,13 @@ impl<T: Config> Pallet<T> {
             let era_duration = (now_as_millis_u64 - active_era_start).saturated_into::<u64>();
             let staked = Self::eras_total_stake(&active_era.index);
             let issuance = T::Currency::total_issuance();
-            let (validator_payout, remainder) = T::EraPayout::era_payout(
-                staked,
-                issuance,
-                era_duration,
-                T::MaxVariableInflationTotalIssuance::get(),
-                T::FixedYearlyReward::get(),
-            );
+            let (validator_payout, remainder) =
+                T::EraPayout::era_payout(staked, issuance, era_duration);
 
             // Polymesh change
             // -----------------------------------------------------------------
             // Schedule rewards
-            let next_block_number = <frame_system::Pallet<T>>::block_number() + 1u32.into();
-            for (index, validator_id) in T::SessionInterface::validators().into_iter().enumerate() {
-                let schedule_block_number =
-                    next_block_number + index.saturated_into::<T::BlockNumber>();
-                match T::RewardScheduler::schedule(
-                    DispatchTime::At(schedule_block_number),
-                    None,
-                    HIGHEST_PRIORITY,
-                    RawOrigin::Root.into(),
-                    Call::<T>::payout_stakers_by_system {
-                        validator_stash: validator_id.clone(),
-                        era: active_era.index,
-                    }.into()
-                ) {
-                    Ok(_) => log!(
-                        info,
-                        "💸 Rewards are successfully scheduled for validator id: {:?} at block number: {:?}",
-                        &validator_id,
-                        schedule_block_number,
-                    ),
-                    Err(error) => {
-                        log!(
-                            error,
-                            "⛔ Detected error in scheduling the reward payment: {:?}",
-                            error
-                        );
-                        Self::deposit_event(Event::<T>::RewardPaymentSchedulingInterrupted {
-                            account_id: validator_id,
-                            era: active_era.index,
-                            error
-                        });
-                    }
-                }
-            }
+            Self::schedule_payouts(&active_era);
             // -----------------------------------------------------------------
 
             Self::deposit_event(Event::<T>::EraPaid {
@@ -903,9 +862,7 @@ impl<T: Config> Pallet<T> {
                 }
             } else if Validators::<T>::contains_key(&voter) {
                 validators_seen.saturating_inc();
-                if Self::is_validator_compliant(&voter)
-                    && Self::is_validator_active_balance_valid(&voter)
-                {
+                if Self::is_validator_compliant(&voter) {
                     // if this voter is a validator:
                     let self_vote = (
                         voter.clone(),
@@ -980,9 +937,7 @@ impl<T: Config> Pallet<T> {
 
             if Validators::<T>::contains_key(&target) {
                 validators_seen.saturating_inc();
-                if Self::is_validator_compliant(&target)
-                    && Self::is_validator_active_balance_valid(&target)
-                {
+                if Self::is_validator_compliant(&target) {
                     all_targets.push(target);
                 }
             }
@@ -1096,62 +1051,6 @@ impl<T: Config> Pallet<T> {
         );
     }
 
-    // Polymesh change
-    // -----------------------------------------------------------------
-
-    /// Returns `true` if active balance is above [`MinValidatorBond`]. Otherwise, returns `false`.
-    pub(crate) fn is_validator_active_balance_valid(who: &T::AccountId) -> bool {
-        if let Some(controller) = Self::bonded(&who) {
-            if let Some(ledger) = Self::ledger(&controller) {
-                return ledger.active >= MinValidatorBond::<T>::get();
-            }
-        }
-        false
-    }
-
-    /// Returns `true` if `stash` has a valid cdd claim and is permissioned. Otherwise, returns `false`.
-    pub(crate) fn is_validator_compliant(stash: &T::AccountId) -> bool {
-        pallet_identity::Pallet::<T>::get_identity(stash).map_or(false, |id| {
-            pallet_identity::Pallet::<T>::has_valid_cdd(id)
-                && Self::permissioned_identity(id).is_some()
-        })
-    }
-
-    /// Returns `true` if `who` has a valid cdd claim. Otherwise, returns `false`.
-    pub(crate) fn is_nominator_compliant(who: &T::AccountId) -> bool {
-        pallet_identity::Pallet::<T>::get_identity(who)
-            .map_or(false, |id| pallet_identity::Pallet::<T>::has_valid_cdd(id))
-    }
-
-    pub(crate) fn get_bonding_duration_period() -> u64 {
-        (T::SessionsPerEra::get()  * T::BondingDuration::get()) as u64 // total session
-            * T::EpochDuration::get() // session length
-            * T::ExpectedBlockTime::get().saturated_into::<u64>()
-    }
-
-    /// Decrease the running count of validators by 1 for the stash identity.
-    pub(crate) fn release_running_validator(stash: &T::AccountId) {
-        if !<Validators<T>>::contains_key(stash) {
-            return;
-        }
-
-        if let Some(did) = pallet_identity::Pallet::<T>::get_identity(stash) {
-            PermissionedIdentity::<T>::mutate(&did, |preferences| {
-                if let Some(p) = preferences {
-                    if p.running_count > 0 {
-                        p.running_count -= 1;
-                    }
-                }
-            });
-            pallet_identity::Pallet::<T>::remove_account_key_ref_count(&stash);
-        }
-    }
-
-    /// Returns the maximum number of validators per identiy
-    pub fn maximum_number_of_validators_per_identity() -> u32 {
-        (T::MaxValidatorPerIdentity::get() * Self::validator_count()).max(1)
-    }
-
     pub fn unbond_balance(
         controller_account: T::AccountId,
         ledger: &mut StakingLedger<T>,
@@ -1219,40 +1118,6 @@ impl<T: Config> Pallet<T> {
 
         Ok(())
     }
-
-    pub(crate) fn base_chill_from_governance(
-        origin: T::RuntimeOrigin,
-        identity: IdentityId,
-        stash_keys: Vec<T::AccountId>,
-    ) -> DispatchResult {
-        T::AdminOrigin::ensure_origin(origin)?;
-
-        for key in &stash_keys {
-            let key_did = pallet_identity::Pallet::<T>::get_identity(&key);
-            // Checks if the stash key identity is the same as the identity given.
-            ensure!(key_did == Some(identity), Error::<T>::NotStash);
-            // Checks if the key is a validator if not returns an error.
-            ensure!(
-                <Validators<T>>::contains_key(&key),
-                Error::<T>::ValidatorNotFound
-            );
-        }
-
-        for key in stash_keys {
-            Self::chill_stash(&key);
-        }
-
-        // Change identity status to be Non-Permissioned
-        PermissionedIdentity::<T>::remove(&identity);
-
-        Self::deposit_event(Event::<T>::PermissionedIdentityRemoved {
-            governance_councill_did: GC_DID,
-            validators_identity: identity,
-        });
-        Ok(())
-    }
-
-    // -----------------------------------------------------------------
 }
 
 impl<T: Config> Pallet<T> {
